+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Version 6.1.10 (18 of January, 2019)
+- Updated `RewardedVideoSampleActivity` bridge to support [*MoPub SDK* v5.x.x](https://developers.mopub.com/docs/android/initialization/)
+- Fixed crash for banner/interstitial `getInstance` calls when android API level < 21: in this case method returns `null`. *Make sure your app code is aware of that*
+- `new LoopMeInterstitial()` and `new LoopMeBanner()` constructor calls are no longer valid: use respective `getInstance` methods
+- Fixed "Problem displaying media file" for VAST ads      
+- other minor bugfixes
+
 ## Version 6.1.9 (07 of December, 2018)
 - Minor gradle bugfixes